WIP: Add line-wrapping functions for formatted_string
DO NOT SUBMIT: add unit tests for these methods.

Crawl is lousy with line-wrapping and string-breaking functions; here's
a few more. These additions permit line-breaking a formatted_string to
fit within a given width and height on the terminal, and determine the
maximum width and height of a multiline string.