Unify some text widget text wrapping code
This change makes use of the new functions for wrapping formatted_text.
Previously, the text widget would pre-wrap the entire text widget
contents into a vector of formatted_strings, one per line. When
rendering, the visible subsection of these would then be displayed. Now,
we use the same breakpointing and slicing algorithm that's used for
local tiles to determine a visible slice of the original string, and
wrap this slice before rendering.

This is arguably more complex and less efficient in the common case,
but it's much more efficient when the text widget contents is extremely
large. In this case, with the addition of a follow up commit that adds
support for nested formatted_string colours, pre-wrapping can cause
visible lag.